+++ docs/release-notes.md
@@ -53,6 +53,9 @@ Example usage in html file: `<link rel="shortcut icon" href="#{assets['img/logo/
 * [1198433](https://bugzilla.redhat.com/show_bug.cgi?id=1198433) - Replace Seam Text with CommonMark Markdown
     * User text on the home page and project "about" pages will now be rendered as CommonMark.
     * Existing Seam Text will be migrated to CommonMark where possible.
+* [1204982](https://bugzilla.redhat.com/show_bug.cgi?id=1204982) - Documentation update for zanata.org/help + readthedocs
+* [1211849](https://bugzilla.redhat.com/show_bug.cgi?id=1211849) - Project maintainer can change project/version slug
+* [1082840](https://bugzilla.redhat.com/show_bug.cgi?id=1082840) - Project maintainer can delete a project or project version
 
 ## 3.6.3